pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "azure-cognitiveservices-speech>=1.49.0",
    "fastapi>=0.135.3",
    "openai>=2.31.0",
    "opencv-python-headless>=4.13.0.92",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.1",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "requests>=2.33.1",
    "uvicorn[standard]>=0.44.0",
]
